fixray-streamlit-app/requirements.txt
streamlit
Pillow
numpy
torch
 torchvision
scikit-learn
matplotlib
opencv-python
pytest